rename particle_tools.py to lagrangian_walker.py and set version=2.0.0
